style(log): use sys/types.h to define data types.

+++ include/log/log_output.h
@@ -10,12 +10,13 @@
 #ifndef LOG_OUTPUT_H
 #define LOG_OUTPUT_H
 #include <stdio.h>
+#include <sys/types.h>
 
 extern FILE* log_file;
 // open a log file
 int InitLogFile(char* log_root_path);
 // main method
-void Logging(unsigned short is_access, char* resource_path, unsigned short return_code, unsigned short method, unsigned short log_mode);
+void Logging(u_int8_t is_access, char* resource_path, u_int16_t return_code, u_int8_t method, u_int8_t log_mode);
 // close log file
 int EndLogFile();
